# lib/swagger/doc/internal.yaml
swagger: '2.0'
info:
  description: Helo Whatsapp API documentation
  version: 1.0.0
  title: Helo Whatsapp
  termsOfService: ''
  contact:
    name: vivaconnect support
    email: support@vivaconnect.co
...
schemes:
  - http
paths:
...
  /templates:
    get:
      isPublic: false
      description: |-
        This api returns list of templates based on filters
        following are the list of template status :
        Approved
        DELETE_PENDING
        Request Initiated
        SUBMITTED
        Rejected
        REQUESTED
        PENDING
      summary: Get template list
      tags:
        - templates
      operationId: Gettemplatelist
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: messageTemplateStatusId
          in: query
          required: false
          type: string
          description: ''
      responses:
        '200':
          description: List of all Templates for waba account
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: array
            example:
              code: 2000
              msg: Success
              data:
                - messageTemplateId: account_update_1
                  TemplateName: account_update_1
                  type: Standard Message Template
                  categoryName: Account Update
                  statusName: Approved
                  languageName: English
                  mediaType: ''
                - messageTemplateId: account_update_2
                  TemplateName: account_update_2
                  type: Standard Message Template
                  categoryName: Account Update
                  statusName: Approved
                  languageName: English
                  mediaType: ''
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
    post:
      isPublic: false
      description: |-
        API to add or update template
        To add template do not pass templateID to update template pass template ID along with parameters to update
      summary: Add update template
      tags:
        - templates
      operationId: Addupdatetemplate
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/AddupdatetemplateRequest'
      responses:
        '200':
          description: Returns Status of template info completion
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data:
                  mediaTemplateComplete: true
              - code: 3022
                msg: Maximum number of template allowed for this account has exausted.
                data: {}
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  /templates/count:
    get:
      description: This api returns templates count
      summary: Get template count
      tags:
        - templates
      operationId: getTemplateCount
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
      responses:
        '200':
          description: This api returns templates count
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: array
            example:
              code: 2000
              msg: Success
              data:
                statusCount:
                  - statusName: Rejected
                    templateCount: 0
                  - statusName: Approved
                    templateCount: 0
                  - statusName: Requested
                    templateCount: 0
                  - statusName: Submit-Failed
                    templateCount: 0
                  - statusName: Partially Approved
                    templateCount: 0
                  - statusName: Denied
                    templateCount: 0
                  - statusName: Deleted
                    templateCount: 0
                  - statusName: Pending
                    templateCount: 0
                  - statusName: Submitted
                    templateCount: 0
                  - statusName: Complete
                    templateCount: 0
                  - statusName: Incomplete
                    templateCount: 0
                allocatedTemplateCount: 0
                usedTemplateCount: 0
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  '/templates/{templateId}':
    get:
      summary: Get template Info
      description: Use this API to get information of a template
      tags:
        - templates
      operationId: GettemplateInfo
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: templateId
          in: path
          required: true
          type: string
          description: Id of Template
      responses:
        '200':
          description: Returns info of a template
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data:
                  - messageTemplateId: account_update_1
                    wabaInformationId: f742d535-4161-4a11-b99f-0c97154b720d
                    templateName: account_update_1
                    type: standard message template
                    bodyText: 'This is to remind you that {{1}} is due by {{2}}.'
                    headerText: ''
                    footerText: ''
                    mediaType: ''
                    categoryName: Account Update
                    statusName: Approved
                    languageName: English
                    secondLanguageRequired: false
                    secondLanguageName: null
                    secondlanguageBodyText: null
                    headerType: null
                    buttonType: null
                    buttonData: null
                    mediaTemplateComplete: false
              - code: 3000
                msg: No record found.
                data: null
                error: {}
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
    delete:
      summary: Delete Template
      description: Use this API to delete a template
      tags:
        - templates
      operationId: deleteTemplate
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: templateId
          in: path
          required: true
          type: string
          description: Id of Template
      responses:
        '200':
          description: Delete a template
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data: {}
              - code: 3000
                msg: No record found.
                data: null
                error: {}
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  '/templates/{templateId}/status':
    patch:
      summary: API to update template status
      description: API to template status
      tags:
        - templates
      operationId: updateTemplateStatus
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: Internal Auth Token
        - name: templateId
          in: path
          required: true
          type: string
          description: Id of Template
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/UpdateTemplateStatusRequest'
      responses:
        '200':
          description: verification success message
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data: {}
                error: null
              - code: 3048
                msg: Status cannot be changed
                data: null
                error:
                  details: cannot change to a new requested status from this current status
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              - code: 4000
                msg: Invalid request
                data: null
                error:
                  - Invalid request
              - code: 3026
                msg: Invalid request
                data: null
                error:
                  - TemplateId does not exist for this waba number
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            example:
              - Unauthorized
              - code: 3017
                msg: Invalid verification code
                data: null
                error: {}
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  '/templates/{templateId}/submit':
    post:
      description: API to send template for approval
      summary: API to send template for approval
      tags:
        - templates
      operationId: sendTemplateForApproval
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: templateId
          in: path
          required: true
          type: string
          description: Id of Template
      responses:
        '200':
          description: API to send template for approval
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data: {}
                error: null
              - code: 3053
                msg: Template not found
                data: null
                error: {}
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
                - Please ensure that the template is in complete status
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  '/templates/{templateId}/submit/{evaluationResponse}':
    patch:
      summary: API to evaluate template
      description: API to evaluate template
      tags:
        - templates
      operationId: sendTemplateForEvaluaion
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: Internal Auth Token
        - name: templateId
          in: path
          required: true
          type: string
          description: Id of Template
        - name: Body
          in: body
          description: ''
          schema:
            $ref: '#/definitions/templateEvaluationRequest'
        - name: evaluationResponse
          in: path
          required: true
          type: string
          description: Evaluation Response of Template
      responses:
        '200':
          description: Template evaluation message
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data: {}
                error: null
              - code: 3053
                msg: Evaluation cannot be proceeded as the approval or rejection from support person is not received.
                data: null
                error: {}
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              - code: 4000
                msg: Invalid request
                data: null
                error:
                  - Invalid request
              - code: 3026
                msg: Invalid request
                data: null
                error:
                  - TemplateId does not exist for this waba number
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            example:
              - Unauthorized
              - code: 3017
                msg: Invalid verification code
                data: null
                error: {}
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  /templates/languages:
    get:
      description: API to get list of languages available with helo-whatsapp
      summary: get template languages
      tags:
        - templates
      operationId: templatelanguagesdropdown
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
      responses:
        '200':
          description: List of all languages
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: array
            example:
              code: 2000
              msg: Success
              data:
                - messageTemplateLanguageId: 5de770f6-3fc0-4eca-b9aa-da1653be9ac7
                  languageName: Select Language
                - messageTemplateLanguageId: 57aa7635-5717-4397-8eb6-4799ef3bec05
                  languageName: English
                - messageTemplateLanguageId: cde192b7-c60f-4d2d-8573-dd5af7c9379b
                  languageName: Marathi
                - messageTemplateLanguageId: 8e50adb9-d12f-43f2-968b-7fa19e1fb761
                  languageName: Gujarat
                - messageTemplateLanguageId: b3c06bd2-7460-4efa-9c72-388e3ecd13da
                  languageName: Punjabi
                - messageTemplateLanguageId: 21c843cb-37de-4b88-914f-82620505276b
                  languageName: Bengali
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  /templates/types:
    get:
      description: API to get list of template types available with helo-whatsapp
      summary: Get template types
      tags:
        - templates
      operationId: templatetypesdropdown
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
      responses:
        '200':
          description: List of all languages
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: array
            example:
              code: 2000
              msg: Success
              data:
                - templateType: Standard Message Template
                - templateType: Media Message Template
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  /templates/categories:
    get:
      description: API to get list of template categories available with helo-whatsapp
      summary: get templates categories
      tags:
        - templates
      operationId: templatescategoriesdropdown
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
      responses:
        '200':
          description: List of all languages
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: array
            example:
              code: 2000
              msg: Success
              data:
                - messageTemplateCategoryId: b8203a31-e439-4ea4-a270-bd211317d3ff
                  categoryName: Account Update
                - messageTemplateCategoryId: 957e4375-3297-42c6-a9a8-6a57fd8e5045
                  categoryName: Alert Update
                - messageTemplateCategoryId: 15b0c326-2bee-401a-8568-a6e27d8f0af0
                  categoryName: Appointment Update
                - messageTemplateCategoryId: 6feb234a-4513-486a-8f4a-977dfb6e1597
                  categoryName: Issue Resolution
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
...
  /templates/buttonType:
    get:
      summary: get buttonType
      description: get list of all available button types in helo-whatsapp
      tags:
        - templates
      operationId: getbuttonType
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
      responses:
        '200':
          description: List of all available button types
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: array
            example:
              code: 2000
              msg: Success
              data:
                - buttonType: Call To Action
                - buttonType: Quick Reply
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
...
definitions:
...
  UpdateTemplateStatusRequest:
    title: updateTemplateStatusRequest
    example:
      firstLocalizationNewStatusId: c71a8387-80e0-468b-9ee3-abb5ec328176
      firstLocalizationOldStatusId: 1cc8cc1f-282a-4431-8618-43effb1ef7c0
      secondLocalizationNewStatusId: 1d9d14ca-d3ec-4bea-b3de-05fcb8ceabd9
      secondLocalizationOldStatusId: b4414c85-5f80-4e8d-98bc-44bbc05b14b1
      firstLocalizationRejectionReason: MARKETING STUFF
      secondLocalizationRejectionReason: MARKETING STUFF
    type: object
    properties:
      firstLocalizationNewStatusId:
        type: string
      firstLocalizationOldStatusId:
        type: string
      firstLocalizationRejectionReason:
        type: string
      secondLocalizationNewStatusId:
        type: string
      secondLocalizationOldStatusId:
        type: string
      secondLocalizationRejectionReason:
        type: string
    required:
      - firstLocalizationNewStatusId
      - firstLocalizationOldStatusId
...
  templateEvaluationRequest:
    title: templateEvaluationRequest
    example:
      firstLocalizationRejectionReason: null
      secondLocalizationRejectionReason: null
    type: object
    properties:
      firstLocalizationRejectionReason:
        type: string
      secondLocalizationRejectionReason:
        type: string